+++ R/diagmeta-internal.R
@@ -194,3 +194,10 @@ trapz <- function (x, y) {
   idx = 2:length(x)
   return(as.double((x[idx] - x[idx - 1]) %*% (y[idx] + y[idx - 1])) / 2)
 }
+
+invert <- function(x, direction, min, max) {
+  if (direction == "increasing")
+    return(x)
+  else
+    return(min + max - x)
+}